// src/api/generated/model/createQuestionGroupRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { CreateQuestionGroupHeaderRequest } from "./createQuestionGroupHeaderRequest";

export interface CreateQuestionGroupRequest {
  id?: string;
  /** @minLength 1 */
  name: string;
  examTypeId: string;
  examSectionId: string;
  questionGroupTypeId: string;
  maximumScore?: number;
  durationInSeconds?: number;
  headers?: CreateQuestionGroupHeaderRequest[];
}