// src/api/generated/model/createQuestionGroupHeaderRequest.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */
import type { CreateQuestionGroupHeaderRequestMediaType } from "./createQuestionGroupHeaderRequestMediaType";

export interface CreateQuestionGroupHeaderRequest {
  id: string;
  orderNumber?: number;
  mediaType?: CreateQuestionGroupHeaderRequestMediaType;
  /** @minLength 1 */
  content: string;
}