// src/api/generated/model/createQuestionGroupHeaderRequestMediaType.ts
/**
 * Generated by orval v8.0.1 🍺
 * Do not edit manually.
 * Measuring Parent API
 * API documentation for Measuring Parent application
 * OpenAPI spec version: 1.0.0
 */

export type CreateQuestionGroupHeaderRequestMediaType =
  (typeof CreateQuestionGroupHeaderRequestMediaType)[keyof typeof CreateQuestionGroupHeaderRequestMediaType];

export const CreateQuestionGroupHeaderRequestMediaType = {
  IMAGE: "IMAGE",
  VIDEO: "VIDEO",
  AUDIO: "AUDIO",
  DOCUMENT: "DOCUMENT",
  PDF: "PDF",
  TEXT: "TEXT",
  ARCHIVE: "ARCHIVE",
  OTHER: "OTHER",
} as const;